Remove RenderWidget.hh
It is actually not needed at all since we never #include
it anywhere.  Just move the class definition into the .cc
file, and get rid of unnecessary infrastructure.
